# openapi.yaml
openapi: 3.0.3
info:
  title: RefOnRecord – Resume Verifier API
  description: >
    API for managing resume projects, work experiences, and verification requests.
    Verifiers can be emailed to confirm or reject employment history.
  version: "1.0.0"
  contact:
    name: PWP Spring 2026 Group
    email: Sillah.Babar@student.oulu.fi
servers:
  - url: https://refonrecord.onrender.com/api
    description: Production
  - url: http://localhost:5000/api
    description: Local development
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
  responses:
    BadRequest:
      description: Invalid request body or validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, or expired bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated user is not the owner of this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Duplicate resource (e.g. email or username already in use)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        user_id:
          type: integer
        email:
          type: string
          format: email
        username:
          type: string
        phone_number:
          type: string
          nullable: true
        tier:
          type: string
          enum: [normal, premium]
        created_at:
          type: string
          format: date-time
        links:
          type: object
          properties:
            self:
              type: string
            projects:
              type: string
    Project:
      type: object
      properties:
        project_id:
          type: integer
        user_id:
          type: integer
        project_name:
          type: string
        template_style:
          type: string
          enum: [classic, modern, minimal]
        phone_number:
          type: string
          nullable: true
        linkedin_url:
          type: string
          nullable: true
        github_url:
          type: string
          nullable: true
        personal_website:
          type: string
          nullable: true
        current_company:
          type: string
          nullable: true
        is_employed:
          type: boolean
        education_details:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        links:
          type: object
          properties:
            self:
              type: string
            owner:
              type: string
            experiences:
              type: string
            shares:
              type: string
...
paths:
...
  /users/:
    post:
      tags: [Users]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, username, password]
              additionalProperties: false
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                username:
                  type: string
                  minLength: 3
                  maxLength: 100
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                phone_number:
                  type: string
                  maxLength: 20
                tier:
                  type: string
                  enum: [normal, premium]
            example:
              email: user@example.com
              username: johndoe
              password: securepass1
      responses:
        "201":
          description: User created
          headers:
            Location:
              schema:
                type: string
              description: URL of the created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Email or username already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{user_id}/:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Users]
      summary: Get a user profile (owner only)

      security:
        - BearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Users]
      summary: Update a user profile (owner only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              additionalProperties: false
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                username:
                  type: string
                  minLength: 3
                  maxLength: 100
                phone_number:
                  type: string
                  maxLength: 20
                tier:
                  type: string
                  enum: [normal, premium]
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Users]
      summary: Delete a user account and all associated data (owner only)
      security:
        - BearerAuth: []
      responses:
        "204":
          description: User deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Projects ─────────────────────────────────────────────────────────────────
  /users/{user_id}/projects/:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Projects]
      summary: List all projects for a user (owner only, cached 300s)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Projects]
      summary: Create a new resume project (owner only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_name]
              additionalProperties: false
              properties:
                project_name:
                  type: string
                  minLength: 1
                  maxLength: 200
                template_style:
                  type: string
                  enum: [classic, modern, minimal]
                phone_number:
                  type: string
                  maxLength: 20
                linkedin_url:
                  type: string
                  maxLength: 500
                github_url:
                  type: string
                  maxLength: 500
                personal_website:
                  type: string
                  maxLength: 500
                current_company:
                  type: string
                  maxLength: 200
                is_employed:
                  type: boolean
                education_details:
                  type: string
      responses:
        "201":
          description: Project created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{user_id}/projects/{project_id}/:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Projects]
      summary: Get a single project (owner only, cached 300s)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Projects]
      summary: Update a project (owner only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              additionalProperties: false
              properties:
                project_name:
                  type: string
                  minLength: 1
                  maxLength: 200
                template_style:
                  type: string
                  enum: [classic, modern, minimal]
                phone_number:
                  type: string
                  maxLength: 20
                linkedin_url:
                  type: string
                  maxLength: 500
                github_url:
                  type: string
                  maxLength: 500
                personal_website:
                  type: string
                  maxLength: 500
                current_company:
                  type: string
                  maxLength: 200
                is_employed:
                  type: boolean
                education_details:
                  type: string
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Projects]
      summary: Delete a project and all associated data (owner only)
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Project deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Experiences ──────────────────────────────────────────────────────────────
...